requirements.txt
streamlit>=1.32.0
opencv-python-headless>=4.9.0
Pillow>=10.0.0
numpy>=1.24.0
matplotlib>=3.7.0
scikit-image>=0.21.0
scipy>=1.11.0
